// web/src/kiosk/components/__generated__/ScanStatusDialogQuery.graphql.ts
export type ScanStatusDialogQuery$variables = {
  first: number;
};
export type ScanStatusDialogQuery$data = {
  readonly session: {
    readonly location: {
      readonly periods: {
        readonly edges: ReadonlyArray<{
          readonly node: {
            readonly guestName: string | null | undefined;
            readonly id: string;
            readonly person: {
              readonly firstName: string;
              readonly id: string;
              readonly lastName: string;
            } | null | undefined;
            readonly startTime: number;
          };
        }>;
      };
    };
  };
};
export type ScanStatusDialogQuery = {
  response: ScanStatusDialogQuery$data;
  variables: ScanStatusDialogQuery$variables;
};